requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
faiss-cpu
pypdf